06.recommend/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
jieba>=0.42.0
streamlit>=1.28.0
openai>=1.0.0
google-genai>=1.0.0
